server/requirements.txt
fastapi
uvicorn[standard]
pillow
imagehash
numpy
